+++ HISTORY.md
@@ -4,6 +4,7 @@
   Thanks to Matthias De Smet.
 - CWL: use json input files for passing inputs instead of flattened command
   line arguments. Improves compatibility with multiple runners.
+- Remove usage of lxml for qsignature and qualimap to avoid icu library errors.
 
 ## 1.0.3 (7 May 2017)